Add modularity for mac and linux

+++ .bashrc
@@ -1,9 +1,17 @@
+custominit() 
+{
+    ./scripts/init.sh
+}
+
 compile()
 {
     set -- "${1}" "${@: -1}"
     set -- "${1}" "${2%.cpp}"
     g++ -std=c++17 -Wshadow -Wall -o "${2}.out" "${2}.cpp" -g -fsanitize=address -fsanitize=undefined -D_GLIBCXX_DEBUG
-    rm -r "${2}.out.dSYM"
+    if [ os == "Mac" ]; then
+        echo "Ptweh! What's this, a mac user!?"
+        rm -r "${2}.out.dSYM"
+    fi
     echo Compiled: ${2}.out
     while getopts ":a" opt; do
       case $opt in
@@ -21,10 +29,13 @@ fast()
 {
     set -- "${1%.cpp}"
     g++ -std=c++17 -Wshadow -Wall -o "${1}.out" "${1}.cpp"
-    rm -r "${1}.out.dSYM"
+    if [ os == "Mac" ]; then
+        echo "Ptweh! What's this, a mac user!?"
+        rm -r "${2}.out.dSYM"
+    fi
 }
 
-cowtell()
+cowtel()
 {
     fortune | cowsay
 }
@@ -36,20 +47,26 @@ cowread()
 
 repo()
 {
-    open https://github.com/POeticPotatoes/${1}
-}
-
-submit()
-{
-    set -- "${1%.cpp}"
-    set -- "${1/%/.cpp}"
-    cp $1 ~/Desktop/Work/cpp/submissions/${1}
-    sed -i '' '1s/#include.*/#include <bits\/stdc++.h>/g' ~/Desktop/Work/cpp/submissions/${1}
+    if [ os == "Mac" ]; then
+        echo "Bleaugh! Weird mac syntax..."
+        open https://github.com/POeticPotatoes${1}
+        return
+    fi
+    xdg-open https://github.com/POeticPotatoes/${1}
 }
 
 solution()
 {
     set -- "${1%.cpp}"
-    set -- "${1/%/.cpp}"
-    printf "#include \"/Users/poeticpotato/Desktop/Work/cpp/bits.h\"\nusing namespace std;\n\nint main() {\n    ios_base::sync_with_stdio(false);\n    cin.tie(NULL);\n}" > ${1}
+    printf "#include <bits/stdc++.h>\nusing namespace std;\n\nint main() {\n    ios_base::sync_with_stdio(false);\n    cin.tie(NULL);\n}" > ${1}.cpp
+}
+
+os()
+{
+    STR=$(hostnamectl)
+    if [[ "$STR" == *"Linux"* ]]; then
+        echo "Linux"
+        return
+    fi
+    echo "Mac"
 }
